+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.0.18](https://github.com/zone-eu/wildduck-zonemta-zilter/compare/v1.0.17...v1.0.18) (2024-11-19)
+
+
+### Bug Fixes
+
+* **db-error-return:** ZMS-189-1 ([#42](https://github.com/zone-eu/wildduck-zonemta-zilter/issues/42)) ([07978d0](https://github.com/zone-eu/wildduck-zonemta-zilter/commit/07978d01fb9fd8b84b5894ba2dd91d402f4f377e))
+
+## [1.0.17](https://github.com/zone-eu/wildduck-zonemta-zilter/compare/v1.0.16...v1.0.17) (2024-11-15)
+
+
+### Bug Fixes
+
+* **addrview:** ZMS-189 ([#39](https://github.com/zone-eu/wildduck-zonemta-zilter/issues/39)) ([20c1604](https://github.com/zone-eu/wildduck-zonemta-zilter/commit/20c16045c644c28f0ec6fa67801e34aa1d7f1a84))
+
+## [1.0.16](https://github.com/zone-eu/wildduck-zonemta-zilter/compare/v1.0.15...v1.0.16) (2024-11-14)
+
+
+### Bug Fixes
+
+* **normalization-alias-logging:** ZMS-189 ([#37](https://github.com/zone-eu/wildduck-zonemta-zilter/issues/37)) ([18fce99](https://github.com/zone-eu/wildduck-zonemta-zilter/commit/18fce998cc8584aeed92d52b2237f6dde396e9a1))
+
+## [1.0.15](https://github.com/zone-eu/wildduck-zonemta-zilter/compare/v1.0.14...v1.0.15) (2024-11-07)
+
+
+### Bug Fixes
+
+* **SMTP-message:** ZMS-185 ([#34](https://github.com/zone-eu/wildduck-zonemta-zilter/issues/34)) ([a54bab9](https://github.com/zone-eu/wildduck-zonemta-zilter/commit/a54bab947f150e5c4d70076f51e8ad770cd76ea7))
+
 ## [1.0.14](https://github.com/zone-eu/wildduck-zonemta-zilter/compare/v1.0.13...v1.0.14) (2024-11-07)